fix: fixed the handleClickDelete btn when it clicks it sends an empty array to the backend and gets and error, added conditions

+++ src/customPageComponents/PageComponent.js
@@ -51,13 +51,15 @@ const PageComponent = ({ fetchApiUrl, headerName }) => {
   };
 
   const handleCardDelete = (id) => {
-    handleCardDeleteCall({
-      method: "delete",
-      fetchApiUrl,
-      selectedIds: checkedItemIds.ids.length !== 0 ? checkedItemIds.ids : id,
-      setCheckedItemIds,
-      setIsAuth,
-    });
+    if (checkedItemIds.ids.length > 0 || id.length > 0) {
+      handleCardDeleteCall({
+        method: "delete",
+        fetchApiUrl,
+        selectedIds: checkedItemIds.ids.length !== 0 ? checkedItemIds.ids : id,
+        setCheckedItemIds,
+        setIsAuth,
+      });
+    }
   };
 
   const changePage = ({ selected }) => {
